+++ rocprim/include/rocprim/iterator/zip_iterator.hpp
@@ -195,14 +195,18 @@ public:
         return detail::dereference_iterator_tuple<reference>(iterator_tuple_);
     }
 
-    ROCPRIM_HOST_DEVICE inline
-    pointer operator->() const
+    [[deprecated("This operator is no longer supported and will be removed in the next major "
+                 "release.")]]
+    ROCPRIM_HOST_DEVICE
+    inline pointer
+        operator->() const
     {
         return &(*(*this));
     }
 
-    ROCPRIM_HOST_DEVICE inline
-    reference operator[](difference_type distance) const
+    ROCPRIM_HOST_DEVICE
+    inline reference
+        operator[](difference_type distance) const
     {
         zip_iterator i = (*this) + distance;
         return *i;